fix : Bug on import no-unresolve due to svelte-modals lib. #3858
