+++ play/src/front/Components/MapEditor/ActionPopupOnPersonalAreaWithEntities.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
+    // eslint-disable-next-line import/no-unresolved
     import { closeModal } from "svelte-modals";
-    import PopupGeneric from "../Modal/PopupGeneric.svelte";
+    import PopupGeneric from "../Modal/Popup.svelte";
     import LL from "../../../i18n/i18n-svelte";
 
 
@@ -16,12 +17,12 @@
     <h1 slot="title">{$LL.mapEditor.areaEditor.actionPopupOnPersonalAreaWithEntities.title()}</h1>
     <div slot="content"
          class="tw-whitespace-nowrap">{$LL.mapEditor.areaEditor.actionPopupOnPersonalAreaWithEntities.description()}</div>
-    <div slot="action" class="tw-flex">
+    <svelte:fragment slot="action">
         <button
             on:click={()=>{onDeleteEntities();closeModal()}}>{$LL.mapEditor.areaEditor.actionPopupOnPersonalAreaWithEntities.buttons.remove()}</button>
         <button
             on:click={()=>{onKeepEntities();closeModal()}}>{$LL.mapEditor.areaEditor.actionPopupOnPersonalAreaWithEntities.buttons.keep()}</button>
         <button
             on:click={()=>{onCancel();closeModal()}}>{$LL.mapEditor.areaEditor.actionPopupOnPersonalAreaWithEntities.buttons.cancel()}</button>
-    </div>
+    </svelte:fragment>
 </PopupGeneric>
