+++ play/src/front/Components/Modal/MainModal.svelte
@@ -1,4 +1,5 @@
 <script>
+    // eslint-disable-next-line import/no-unresolved
     import { closeModal, Modals } from "svelte-modals";
 </script>
 
@@ -17,6 +18,6 @@
         bottom: 0;
         right: 0;
         left: 0;
-        background: rgba(0,0,0,0.50)
+        background: rgba(0,0,0,0.50);
     }
 </style>
