+++ play/src/front/Phaser/Game/MapEditor/Tools/AreaEditorTool.ts
@@ -3,6 +3,7 @@ import type { EditMapCommandMessage } from "@workadventure/messages";
 import type { Unsubscriber } from "svelte/store";
 import { get } from "svelte/store";
 import { v4 as uuid } from "uuid";
+// eslint-disable-next-line import/no-unresolved
 import { openModal } from "svelte-modals";
 import {
     mapEditorAreaModeStore,
